accessible/tests/mochitest/test_aria_role_article.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=429666
-->
<head>
  <title>Expose ROLE_DOCUMENT for ARIA landmarks that inherit from document chrome tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="common.js"></script>
  <script type="application/javascript"
          src="role.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      // Test article exposed as document
      testRole("testArticle", ROLE_DOCUMENT);

      var articleAcc = getAccessible("testArticle");
      if (articleAcc)
        is(articleAcc.name, "Test article", "Wrong name for article!");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=429666">Mozilla Bug 429666</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
  <div id="testArticle" role="article" title="Test article">
    <p>This is a paragraph inside the article.</p>
  </div>
</body>
</html>
